requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
dateparser==1.3.0
dj-database-url==3.1.2
django-cors-headers==4.9.0
Django==6.0.3
djangorestframework==3.17.0
idna==3.11
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
regex==2026.2.28
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
openai>=1.0.0
